DS-2841 Remove singular value checks
Removed code block that caused algorithm to throw an error if any
singular values of the square correlation matrix are too small
(eigenvalues would have been more appropriate but it is now irrelevant).
The test for a redundant variable throwing an error has been replaced
with the warning thrown by cor.smooth. This should be investigated and
tidied up later to give informative messages.