DS-2841-2 Handle duplicate rownames in score output
Duplicate rownames in scores were not handled properly before when
subsets occur since the input data is assumed to be a data.frame where
duplicates are impossible. However, a numeri matrix can handle duplicate
rownames. In the case where duplicates are possible the output defers to
the subset argument to determine which values are NA in the score output
instead of matching the rownames. Unit tests added.