DS-2841 Catch edge case in rowname check of scores
The rowname check for the alignment of observations can still fail when
there are no duplciated rownames (they are unique) but one of the
rownames is an empty string. This can occur in Text data where cases are
unique with one person giving an empty string. This commit fixes this
and unit test added.