DS-3022 Do not require original.data to print PCA
TextPCA return list object size is dominated by the original.data
Removing this element saves a lot of object size for large text analysis
inputs. This commit slightly tweaks the code to not require
original.data to determine the number of eigenvalues to compute for the
footer and will inspect the eigenvalue vector x$values if the original
data is unavailable.