DS-3375 Restore rowSums/colSums to avoid error with spans
When a banner table is used as an input to Correspondence Analysis, NET rows/cols
are removed in a call to flipTables::TidyTabular data. However, the span attribute
is not updated to reflect this removal. The subsequent call to verbs::SumRows will
throw an error since it attempts to tidy the labels based off the span attribute.
This commit avoids this by not using SumRows but instead using rowSums instead.